requirements.txt
streamlit
python-dotenv
google-generativeai
Pillow
opencv-python-headless
ultralytics